SDK-1295: Added condition to check ENV setting and config

+++ bin/spryker-sdk.sh
@@ -120,6 +120,10 @@ case $MODE in
         docker-compose -f "${SDK_DIR}/docker-compose.yml" run --rm spryker-sdk "$ARGUMENTS"
     else
         install_composer
+        if [[ ! -f "${SDK_DIR}/docker-compose.dev.yml" ]]; then
+           echo "SPRYKER_SDK_ENV=$SPRYKER_SDK_ENV environment is not available for installer. Remove or set ENV variable \"export SPRYKER_SDK_ENV=prod\""
+           exit 1
+        fi;
         docker-compose -f "${SDK_DIR}/docker-compose.dev.yml" run --rm -e XDEBUG_MODE=off spryker-sdk "$ARGUMENTS"
     fi
   ;;
